requirements.txt
influxdb3-python==0.10.0
python-dotenv==1.0.0
pandas>=2.0.3,<2.1.0
numpy>=1.24.0,<1.25.0
scikit-learn>=1.3.0,<1.4.0
matplotlib>=3.7.1,<3.8.0
qdrant-client==1.6.4